# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "A service that permits to generate unique IUV codes for each creditor institution.\n\n**STANDARD ERRORS:**\nNAME | CODE | DESCRIPTION\n- | - | -\n**IUVG-1000**  | *GENERIC_ERROR* | Error while generating IUV.  *...error description...*\n**IUVG-1001**  | *VALIDATION_ERROR* | Error while validating IUV generation request.  *...error description...*\n**IUVG-1101**  | *GENERATION_MAX_RETRIES_REACHED* | Error while generating IUV. Reached the maximum number of internal automatic attempts  [*...content...*] to generate a unique IUV for creditor institution  [*...content...*]. Please, try again later.\n**IUVG-1102**  | *GENERATION_AUXDIGIT_ALGORITHM_INVALID_PATTERN* | Error while generating IUV. The IUV code's section generated by timestamp in milliseconds is not correctly composited of 13 digits.\n",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "IUV Generator",
    "version" : "0.2.4"
  },
  "servers" : [ {
    "url" : "${host}",
    "description" : "Generated server URL"
  }, {
    "url" : "http://localhost:8080",
    "description" : "Local server URL"
  } ],
  "tags" : [ {
    "description" : "Application info APIs",
    "name" : "Home"
  }, {
    "description" : "IUV Generation APIs",
    "name" : "Generation"
  } ],
  "paths" : {
    "/info" : {
      "get" : {
        "operationId" : "healthCheck",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AppInfoResponse"
                }
              }
            },
            "description" : "OK."
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "Return OK if application is started",
        "tags" : [ "Home" ]
      }
    },
    "/organizations/{organization-fiscal-code}/iuv" : {
      "post" : {
        "description" : "**Description:**  \nGenerate a new unique IUV code without aux-digit at the start.  \n\n**API properties:**\nProperty | Value\n- | -\nInternal | Y\nExternal | N\nSynchronous | SYNC\nAuthorization | NONE\nAuthentication | APIKEY\nTPS | 1.0/sec\nIdempotency | N\nStateless | Y\nRead/Write Intense | Read and Write\nCacheable | N\n",
        "operationId" : "generateIUV",
        "parameters" : [ {
          "description" : "The fiscal code of the creditor institution",
          "example" : 77777777777,
          "in" : "path",
          "name" : "organization-fiscal-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IUVGenerationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IUVGenerationResponse"
                }
              }
            },
            "description" : "Created."
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "Generate new IUV for organization",
        "tags" : [ "Generation" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "AppInfoResponse" : {
        "type" : "object",
        "properties" : {
          "db" : {
            "type" : "string"
          },
          "environment" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        }
      },
      "IUVGenerationRequest" : {
        "type" : "object",
        "properties" : {
          "auxDigit" : {
            "type" : "integer",
            "description" : "The value of the AUX digit related to the IUV.",
            "format" : "int32",
            "example" : 3
          },
          "segregationCode" : {
            "type" : "integer",
            "description" : "The value of the segregation related to the IUV.",
            "format" : "int32",
            "example" : 48
          }
        },
        "required" : [ "auxDigit", "segregationCode" ]
      },
      "IUVGenerationResponse" : {
        "type" : "object",
        "properties" : {
          "iuv" : {
            "type" : "string",
            "description" : "The unique IUV generated from computation for the creditor institution.",
            "example" : "584756566583"
          }
        },
        "required" : [ "iuv" ]
      }
    },
    "securitySchemes" : {
      "ApiKey" : {
        "description" : "The API key to access this function app.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}