# src/main/java/it/gov/pagopa/iuvgenerator/controller/model/AppInfoResponse.java
package it.gov.pagopa.iuvgenerator.controller.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.*;

@Data
@Builder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor(access = AccessLevel.PRIVATE)
@ToString
@JsonIgnoreProperties(ignoreUnknown = true)
public class AppInfoResponse {

    private String name;
    private String version;
    private String environment;
    private String db;
}